test: refactor some cases: compare intersections of exptecteds & actuals
A helper function was added to `assert.are_same` on the *intersections* of given
tables. This allows some test cases to only reason about those keys of the
analyzed table, which the test specifies / is aware of. This allows adding new
fields to a table returned by unit under test without breaking existing tests so
often. This also allows many cases to contain only a single assert, which is
also considered a good practice.